Move external `VersionReq` deserialization into function to prevent breaking internal Manifest storage, ensure that `default_version` is used when retrieving an unknown world from a `Manifest`
